+++ figlet4s-effects/src/test/scala/com/colofabrix/scala/figlet4s/either/EitherFiglet4sSpecs.scala
@@ -1,5 +1,6 @@
 package com.colofabrix.scala.figlet4s.either
 
+import cats.scalatest._
 import cats.effect._
 import cats.implicits._
 import com.colofabrix.scala.figlet4s._
@@ -11,12 +12,12 @@ import org.scalatest.flatspec._
 import org.scalatest.matchers.should._
 import scala.concurrent._
 
-class EitherFiglet4sSpecs extends AnyFlatSpec with Matchers with Figlet4sMatchers {
+class EitherFiglet4sSpecs extends AnyFlatSpec with Matchers with Figlet4sMatchers with EitherMatchers with EitherValues {
 
   //  Rendering  //
 
   "Rendering APIs" should "render a default text using the \"standard\" font" in {
-    val test =
+    val computed =
       for {
         options <- standardBuilder.options
       } yield {
@@ -25,18 +26,21 @@ class EitherFiglet4sSpecs extends AnyFlatSpec with Matchers with Figlet4sMatcher
         computed should lookLike(expected)
       }
 
-    test.unsafeGet
+    computed should be(right)
   }
 
   //  Internal fonts  //
 
   "Internal Fonts API" should "return the list of internal fonts containing at least the \"standard\" font" in {
-    Figlet4s.internalFonts.unsafeGet should contain(Figlet4sClient.defaultFont)
+    val computed = Figlet4s.internalFonts
+
+    computed should be(right)
+    computed.value should contain(Figlet4sClient.defaultFont)
   }
 
   it should "load all internal fonts successfully" in {
-    val fonts = Figlet4s.internalFonts.unsafeGet
-    fonts.map(Figlet4s.loadFontInternal(_).unsafeGet)
+    val fonts = Figlet4s.internalFonts.value
+    fonts.map(Figlet4s.loadFontInternal(_).value)
   }
 
   it should "support loading of internal fonts in parallel" in {
@@ -44,13 +48,13 @@ class EitherFiglet4sSpecs extends AnyFlatSpec with Matchers with Figlet4sMatcher
 
     val data = Figlet4s
       .internalFonts
-      .unsafeGet
+      .value
       .toList
       .take(10)
       .flatMap(Seq.fill(10)(_))
 
     val test = data.parTraverse { fontName =>
-      IO(Figlet4s.loadFontInternal(fontName).unsafeGet)
+      IO(Figlet4s.loadFontInternal(fontName).value)
     }
 
     test.unsafeRunSync()
@@ -66,7 +70,8 @@ class EitherFiglet4sSpecs extends AnyFlatSpec with Matchers with Figlet4sMatcher
 
   "Fonts API" should "read a font from the file system" in {
     val cwd = System.getProperty("user.dir")
-    Figlet4s.loadFont(s"$cwd/figlet4s-core/src/main/resources/fonts/standard.flf").unsafeGet
+    val computed = Figlet4s.loadFont(s"$cwd/figlet4s-core/src/main/resources/fonts/standard.flf")
+    computed should be(right)
   }
 
   it should "throw a FigletLoadingError when trying to load a font that doesn't exist" in {